+++ pmd-core/src/main/java/net/sourceforge/pmd/internal/util/IteratorUtil.java
@@ -8,6 +8,7 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Iterator;
 import java.util.List;
+import java.util.ListIterator;
 
 
 /**
@@ -46,4 +47,46 @@ public final class IteratorUtil {
         };
     }
 
+
+    /** Counts the items in this iterator, exhausting it. */
+    public static int count(Iterator<?> it) {
+        int count = 0;
+        while (it.hasNext()) {
+            it.next();
+            count++;
+        }
+        return count;
+    }
+
+
+    public static <T> Iterable<T> asReversed(final List<T> lst) {
+
+        return new Iterable<T>() {
+            @Override
+            public Iterator<T> iterator() {
+                return new Iterator<T>() {
+
+                    ListIterator<T> li = lst.listIterator(lst.size());
+
+
+                    @Override
+                    public boolean hasNext() {
+                        return li.hasPrevious();
+                    }
+
+
+                    @Override
+                    public T next() {
+                        return li.previous();
+                    }
+
+
+                    @Override
+                    public void remove() {
+                        li.remove();
+                    }
+                };
+            }
+        };
+    }
 }
