+++ pmd-ui/src/main/java/net/sourceforge/pmd/util/fxdesigner/util/DesignerIteratorUtil.java
@@ -7,8 +7,6 @@ package net.sourceforge.pmd.util.fxdesigner.util;
 import static net.sourceforge.pmd.internal.util.IteratorUtil.toIterable;
 
 import java.util.Iterator;
-import java.util.List;
-import java.util.ListIterator;
 import java.util.function.Function;
 import java.util.function.Predicate;
 
@@ -25,41 +23,11 @@ public final class DesignerIteratorUtil {
 
     // TODO move that into PMD core with Java 8
 
+
     private DesignerIteratorUtil() {
 
     }
 
-
-    /** Counts the items in this iterator, exhausting it. */
-    public static int count(Iterator<?> it) {
-        int count = 0;
-        while (it.hasNext()) {
-            it.next();
-            count++;
-        }
-        return count;
-    }
-
-
-    public static <T> Iterable<T> asReversed(List<T> lst) {
-
-        return () -> new Iterator<T>() {
-
-            ListIterator<T> li = lst.listIterator(lst.size());
-
-
-            @Override
-            public boolean hasNext() {
-                return li.hasPrevious();
-            }
-
-
-            @Override
-            public T next() {
-                return li.previous();
-            }
-        };
-    }
 
     public static boolean isParent(Node parent, Node child) {
         return any(parentIterator(child, false), p -> parent == p);
